+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## Unreleased
+### Fix
+- Set correct CategoryPath field name for searchParamsFromUrl
+
 ## [v6.0.2] - 2024.12.10
 ### Fix
 - Don't redirect to search result page if search is executed already on result page