+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 ## Unreleased
+### Fix
+- Set correct CategoryPath field name for searchParamsFromUrl
+
 ### Change
 - Update FactFinder logo icon
+
 ## [v6.0.2] - 2024.12.10
 ### Fix
 - Don't redirect to search result page if search is executed already on result page