+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Fix
 - Set correct CategoryPath field name for searchParamsFromUrl
 - Fix Filter off-canvas issue
+- Fix Proxy issue on PDP and for tracking requests
 
 ### Change
 - Update FactFinder logo icon